requirements.txt
tweepy==4.14.0
python-dotenv==1.0.1
openai==1.41.0
polars==1.5.0
streamlit==1.37.1
